Refactors and fully extracts article creation to its own module
