+++ src/components/Dashboard.jsx
@@ -1,12 +1,10 @@
 import React from 'react';
 import ArticleCreationModal from './ArticleCreationModal';
-import PopupMessage from './PopupMessage';
 
 const JournalistDashboard = () => {
   return (
     <>
       <ArticleCreationModal />
-      <PopupMessage />
     </>
   );
 };
